サブピクセル推定の実装

See merge request tech/adaskit/libSGM!23

+++ src/types.hpp
@@ -24,7 +24,7 @@ namespace sgm {
 using feature_type = uint32_t;
 using cost_type = uint8_t;
 using cost_sum_type = uint16_t;
-using output_type = uint8_t;
+using output_type = uint16_t;
 
 }
